Add `VoidResponse`
Remove `ImmediateAction`
